Moves the barcode related functionality out of tasks and into its own location.  Splits up the testing based on that
